+++ src/app/app.component.spec.ts
@@ -2,7 +2,7 @@ import {TestBed} from "@angular/core/testing";
 import {RouterTestingModule} from "@angular/router/testing";
 import {AppComponent} from "./app.component";
 import {of} from "rxjs";
-import {AuthService} from "./services/auth.service";
+import {AuthService} from "./core/services/auth.service";
 import { TranslateService } from '@ngx-translate/core';
 import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
