+++ src/app/core/guards/secure-inner-pages.guard.ts
