+++ src/app/core/services/groups.service.spec.ts
@@ -3,7 +3,7 @@ import {GroupsService} from "./groups.service";
 import {of} from "rxjs";
 import {Timestamp} from "firebase/firestore";
 import {initializeApp} from "firebase/app";
-import {environment} from "../../environments/environment";
+import {environment} from "../../../environments/environment";
 
 initializeApp(environment.firebaseConfig);
