+++ src/app/core/services/users.service.ts
@@ -1,6 +1,6 @@
 import {Injectable} from "@angular/core";
 import {FirestoreService} from "./firestore.service";
-import {AppUser} from "../models/user.model";
+import {AppUser} from "../../models/user.model";
 import {
   addDoc,
   collection,
