+++ src/app/menu/menu.component.ts
@@ -2,7 +2,7 @@ import {CommonModule} from "@angular/common";
 import {Component, OnInit} from "@angular/core";
 import {RouterLink, RouterLinkActive} from "@angular/router";
 import {IonicModule} from "@ionic/angular";
-import {AuthService} from "../services/auth.service";
+import {AuthService} from "../core/services/auth.service";
 import {User} from "firebase/auth";
 
 @Component({
