+++ src/app/modules/group/pages/group-edit/group-edit.page.ts
@@ -2,7 +2,7 @@ import {Component, OnInit} from "@angular/core";
 import {CommonModule} from "@angular/common";
 import {FormsModule} from "@angular/forms";
 import {IonicModule} from "@ionic/angular";
-import {MenuService} from "../../../../services/menu.service";
+import {MenuService} from "../../../../core/services/menu.service";
 
 @Component({
   selector: "app-group-edit",
