+++ src/app/modules/user/pages/user-dashboard/user-dashboard.page.ts
@@ -2,8 +2,8 @@ import {Component, OnInit} from "@angular/core";
 import {CommonModule} from "@angular/common";
 import {FormsModule} from "@angular/forms";
 import {IonicModule, NavController} from "@ionic/angular";
-import {AuthService} from "../../../../services/auth.service";
-import {MenuService} from "../../../../services/menu.service";
+import {AuthService} from "../../../../core/services/auth.service";
+import {MenuService} from "../../../../core/services/menu.service";
 
 @Component({
   selector: "app-user-dashboard",
