+++ src/app/modules/user/pages/user-profile/user-profile.page.ts
@@ -2,9 +2,9 @@ import {Component, OnInit} from "@angular/core";
 import {CommonModule} from "@angular/common";
 import {FormsModule} from "@angular/forms";
 import {IonicModule} from "@ionic/angular";
-import {UsersService} from "../../../../services/users.service";
+import {UsersService} from "../../../../core/services/users.service";
 import {ActivatedRoute} from "@angular/router";
-import {MenuService} from "../../../../services/menu.service";
+import {MenuService} from "../../../../core/services/menu.service";
 
 @Component({
   selector: "app-user-profile",
