+++ src/app/modules/user/pages/user-settings/user-settings.page.ts
@@ -2,7 +2,7 @@ import {Component, OnInit} from "@angular/core";
 import {CommonModule} from "@angular/common";
 import {FormsModule} from "@angular/forms";
 import {IonicModule} from "@ionic/angular";
-import {MenuService} from "../../../../services/menu.service";
+import {MenuService} from "../../../../core/services/menu.service";
 import {TranslateModule} from "@ngx-translate/core";
 import {LanguageSelectorComponent} from "../../components/language-selector/language-selector.component";
